src/requirements.txt
boto3==1.28.34
fuzzywuzzy==0.18.0
matplotlib==3.7.2
numpy==1.25.2
pandas==1.5.3
Pillow==10.2.0
Requests==2.31.0
tweepy==4.10.0
urllib3==1.26.16
wordcloud==1.9.3
